requirements.txt
# ============================================
# CORE FRAMEWORK - Optimized versions
# ============================================
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
python-decouple==3.8

# ============================================
# DATABASE - Updated for compatibility
# ============================================
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================
# AI & ML - CORE ECOSYSTEM (No LangFlow)
# ============================================
openai==1.54.4
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
tiktoken==0.8.0

# PyTorch Ecosystem - UPDATED & ALIGNED
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.46.3
datasets==3.1.0
tokenizers==0.20.3
accelerate==1.1.1
peft==0.13.2
bitsandbytes==0.44.1 ; platform_system != "Windows"

# ============================================
# CACHING & QUEUE
# ============================================
redis==5.2.0
celery==5.4.0
flower==2.0.1

# ============================================
# REAL-TIME COMMUNICATION
# ============================================
websockets==13.1
python-socketio==5.11.4

# ============================================
# NOTIFICATIONS
# ============================================
fastapi-mail==1.4.1
twilio==9.3.7

# ============================================
# FILE HANDLING & STORAGE
# ============================================
python-magic-bin==0.4.14 ; platform_system == "Windows"
python-magic==0.4.27 ; platform_system != "Windows"
pillow==11.0.0
boto3==1.35.62
minio==7.2.10

# ============================================
# MONITORING & LOGGING - Consolidated
# ============================================
loguru==0.7.2
sentry-sdk[fastapi]==2.17.0
prometheus-client==0.21.0

# ============================================
# HTTP CLIENT - Consolidated to httpx
# ============================================
httpx==0.27.2

# ============================================
# DATA PROCESSING
# ============================================
pandas==2.2.3
numpy>=1.26.0,<2.0.0

# ============================================
# TESTING
# ============================================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-xdist==3.6.0  # Parallel test execution

# ============================================
# DEVELOPMENT TOOLS
# ============================================
black==24.10.0
isort==5.13.2
ruff==0.7.4  # Faster alternative to flake8
mypy==1.13.0
pre-commit==4.0.1

# ============================================
# UTILITIES
# ============================================
python-dateutil==2.9.0
pytz==2024.2
click==8.1.7
rich==13.9.4
typer==0.13.1

# ============================================
# CONFIGURATION
# ============================================
python-dotenv==1.0.1
dynaconf==3.2.6

# ============================================
# PERFORMANCE & SECURITY
# ============================================
slowapi==0.1.9
email-validator==2.2.0
gunicorn==23.0.0 ; platform_system != "Windows"
fastapi-health==0.4.0

# ============================================
# WEB SCRAPING - Consolidated
# ============================================
beautifulsoup4==4.12.3
aiofiles==24.1.0
